PR on security semantics
During the Prague side meeting, I took an action item to generate a pull request to describe the security semantic requirements.  It's turned out to be harder than I thought ("don't be evil, and don't use SPUD for evil" didn't quite fly), so I'm throwing up a pull request I know isn't adequate in the hopes it will spur someone to improve it.